// BACKEND/ms-orders/src/shared/enums/AssignmentReportData.enum.ts
export interface AssignmentReportData {
    assignmentId: string;
    assignmentDate: string;
    assignmentStatus: string;
    assignmentNote?: string;
    driverId: string;
    orderId: string;
    orderStatus: string;
    orderTotalAmount?: string;
    orderAddress?: string;
    orderCity?: string;
    orderDepartment?: string;

}